debugnote/requirements.txt
asgiref==3.6.0
astroid==2.15.0
attrs==22.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.1.7
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
drf-yasg==1.21.5
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.0
pathspec==0.11.1
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.17.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
